try moving github-project into github
