+++ src/github/index.ts
@@ -12,4 +12,6 @@ export * from './pull-request-lint';
 
 export * from './task-workflow';
 
+export * from './github-project';
+
 export * as workflows from './workflows-model';
